Fix for: https://issues.jenkins.io/browse/JENKINS-69172 Unable to unshelve change: java.lang.IllegalStateException: Expected 1 instance of hudson.model.User$AllUsers but got 0
When unshelve task executed on a slave node, then Jenkins.getInstance() can not
be used, hence no user found while updating the changelog.

Changelog will get updated once build is successful.
